# frontend/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Annotate Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/task-packages": {
      "get": {
        "summary": "List Packages",
        "operationId": "list_packages_api_v1_task_packages_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackageOut"
                  },
                  "title": "Response List Packages Api V1 Task Packages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Package",
        "operationId": "create_package_api_v1_task_packages_post",
        "parameters": [
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/task-packages/{package_id}/publish": {
      "post": {
        "summary": "Publish Package",
        "operationId": "publish_package_api_v1_task_packages__package_id__publish_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/task-packages/{package_id}/items": {
      "get": {
        "summary": "List Items",
        "operationId": "list_items_api_v1_task_packages__package_id__items_get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ItemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskItemOut"
                  },
                  "title": "Response List Items Api V1 Task Packages  Package Id  Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/annotation-tasks/claim": {
      "post": {
        "summary": "Claim Annotation",
        "operationId": "claim_annotation_api_v1_annotation_tasks_claim_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review-tasks/claim": {
      "post": {
        "summary": "Claim Review",
        "operationId": "claim_review_api_v1_review_tasks_claim_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          },
          {
            "name": "claim_policy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ClaimPolicy"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Claim Policy"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/my-tasks": {
      "get": {
        "summary": "My Tasks",
        "operationId": "my_tasks_api_v1_my_tasks_get",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(annotation|review)$",
              "default": "annotation",
              "title": "Stage"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskItemOut"
                  },
                  "title": "Response My Tasks Api V1 My Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/task-packages/{package_id}/assign": {
      "post": {
        "summary": "Assign Items",
        "operationId": "assign_items_api_v1_task_packages__package_id__assign_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskItemOut"
                  },
                  "title": "Response Assign Items Api V1 Task Packages  Package Id  Assign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/task-items/{item_id}/reclaim": {
      "post": {
        "summary": "Reclaim Item",
        "operationId": "reclaim_item_api_v1_task_items__item_id__reclaim_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReclaimRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignmentRequest": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Item Ids"
          },
          "assignee_id": {
            "type": "string",
            "title": "Assignee Id"
          },
          "stage": {
            "type": "string",
            "pattern": "^(annotation|review)$",
            "title": "Stage"
          }
        },
        "type": "object",
        "required": [
          "item_ids",
          "assignee_id",
          "stage"
        ],
        "title": "AssignmentRequest"
      },
      "ClaimPolicy": {
        "type": "string",
        "enum": [
          "sequential",
          "random"
        ],
        "title": "ClaimPolicy"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "available",
          "annotation_assigned",
          "annotating",
          "review_pending",
          "review_assigned",
          "reviewing",
          "changes_requested",
          "completed",
          "cancelled"
        ],
        "title": "ItemStatus"
      },
...
      "PackageCreate": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "claim_policy": {
            "$ref": "#/components/schemas/ClaimPolicy",
            "default": "sequential"
          },
          "random_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Seed"
          },
          "episode_indices": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Indices"
          },
          "episode_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Start"
          },
          "episode_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode End"
          },
          "member_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Ids"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "dataset_id",
          "title"
        ],
        "title": "PackageCreate"
      },
      "PackageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/PackageStatus"
          },
          "claim_policy": {
            "$ref": "#/components/schemas/ClaimPolicy"
          },
          "random_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Seed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items",
            "default": 0
          },
          "claimed_items": {
            "type": "integer",
            "title": "Claimed Items",
            "default": 0
          },
          "annotated_items": {
            "type": "integer",
            "title": "Annotated Items",
            "default": 0
          },
          "reviewed_items": {
            "type": "integer",
            "title": "Reviewed Items",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "dataset_id",
          "title",
          "description",
          "status",
          "claim_policy",
          "random_seed",
          "created_at"
        ],
        "title": "PackageOut"
      },
      "PackageStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "closed"
        ],
        "title": "PackageStatus"
      },
...
      "QaStatus": {
        "type": "string",
        "enum": [
          "unchecked",
          "passed",
          "rejected"
        ],
        "title": "QaStatus"
      },
...
      "ReclaimRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ReclaimRequest"
      },
...
      "TaskItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "claim_order": {
            "type": "integer",
            "title": "Claim Order"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "annotator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotator Id"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "qa_status": {
            "$ref": "#/components/schemas/QaStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "package_id",
          "episode_id",
          "claim_order",
          "status",
          "annotator_id",
          "reviewer_id",
          "qa_status",
          "updated_at"
        ],
        "title": "TaskItemOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}